requirements.txt
Django==3.2.12
django-cors-headers==3.1.1
django-model-utils==4.0.0
djangorestframework==3.11.0
gitdb==4.0.9
GitPython==3.1.27
